+++ client/src/pages/ImageUpload.js
@@ -1,77 +1,9 @@
-// import React, { useState } from "react";
-// import Button from "@material-ui/core/Button";
-// import { storage, db } from "./firebase";
-// import firebase from "firebase";
-
-// function ImageUpload({ username }) {
-//   const [image, setImage] = useState(null);
-//   const [progress, setProgress] = useState(0);
-//   const [caption, setCaption] = useState("");
-
-//   const handleChange = (e) => {
-//     if (e.target.files[0]) {
-//       setImage(e.target.files[0]);
-//     }
-//   };
-
-//   const handleUpload = () => {
-//     const uploadTask = storage.ref(`images/${image.name}`).put(image);
-
-//     uploadTask.on(
-//       "state_changed",
-//       snapshot => {
-//         const progress = Math.round(
-//           (snapshot.bytesTransferred / snapshot.totalBytes) * 100
-//         )
-//         setProgress(progress);
-//       },
-//       (error) => {
-//         console.log(error);
-//         alert(error.message);
-//       },
-//       () => {
-//         storage
-//           .ref("images")
-//           .child(image.name)
-//           .getDownloadURL()
-//           .then(url => {
-//             db.collection("posts").add({
-//               timestamp: firebase.firestore.FieldValue.serverTimestamp(),
-//               caption: caption,
-//               imageUrl: url,
-//               username: username,
-//             });
-
-//             setProgress(0);
-//             setCaption("");
-//             setImage(null);
-//           });
-//       }
-//     );
-//   };
-
-//   return (
-//     <div>
-//       <progress value={progress} max="100" />
-//       <input
-//         type="text"
-//         placeholder="What's on your mind?"
-//         value={caption}
-//         onChange={(event) => setCaption(event.target.value)}
-//       />
-//       <input type="file" onChange={handleChange} />
-
-//       <Button onclick={handleUpload}>Create Post</Button>
-//     </div>
-//   );
-// }
-
 import React, { useState } from "react";
 import { storage, db } from "./firebase";
 import { makeStyles } from "@material-ui/core/styles";
 import Button from "@material-ui/core/Button";
 import firebase from "firebase";
-import "./ImageUpload.css";
+import "../css/ImageUpload.css";
 import AddAPhotoIcon from "@material-ui/icons/AddAPhoto";
 
 import { Avatar } from "@material-ui/core";
@@ -164,7 +96,7 @@ export default function ImageUpload({ username }) {
     <div className="upload">
       <div className="upload_first">
         <Avatar
-          className="post__avatar"
+          displayName="post__avatar"
           alt="subhampreet"
           src="./images/avatar1.jpg"
           className = {classes.large}
