+++ client/src/pages/Story.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Story.css";
+import "../css/Story.css";
 import { Avatar } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
@@ -26,7 +26,7 @@ function Story() {
     <div className="story">
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="/images/avatar.jpg"
           className={classes.large}
@@ -36,7 +36,7 @@ function Story() {
 
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcQ7X1a5uXND5eV1xt1ihm1RqafYqZ2_iFAWeg&usqp=CAU"
           className={classes.large}
@@ -46,7 +46,7 @@ function Story() {
 
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="https://miro.medium.com/max/2048/0*0fClPmIScV5pTLoE.jpg"
           className={classes.large}
@@ -56,7 +56,7 @@ function Story() {
 
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcRBwgu1A5zgPSvfE83nurkuzNEoXs9DMNr8Ww&usqp=CAU"
           className={classes.large}
@@ -66,7 +66,7 @@ function Story() {
 
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="https://images.unsplash.com/photo-1518806118471-f28b20a1d79d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&w=1000&q=80"
           className={classes.large}
@@ -76,7 +76,7 @@ function Story() {
 
       <div className="story__item">
         <Avatar
-          className="avatar"
+          displayName="avatar"
           alt="subhampreet"
           src="https://i.pinimg.com/736x/0e/c6/6b/0ec66b439eb296c4f69cc261e44a785b.jpg"
           className={classes.large}
