+++ client/src/pages/firebase.js
