+++ client/src/pages/serviceWorker.js
